% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domtools.r
\name{clear}
\alias{clear}
\title{v1.1'
Clear Terminal Function}
\usage{
clear()
}
\description{
This function clears the terminal on Unix systems.
Does not work on Windows.
Run makeActiveBinding("clear", clear, baseenv()) 
to be able to type "clear"
}
\examples{
clear()
}
\keyword{clear}
