% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domtools.r
\name{cls_fun}
\alias{cls_fun}
\title{Clear Terminal Function}
\usage{
cls_fun()
}
\description{
This function clears the terminal on Unix systems.
Does not work on Windows
}
\examples{
cls_fun()
clear
}
\keyword{clear}
