% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domtools.r
\name{colKS}
\alias{colKS}
\title{Kolmogorov-Smirnov Normal Distribution Test for Data Frames}
\usage{
colKS(x, interpret = FALSE)
}
\description{
Function to perform Kolmogorov-Smirnov test for normal distribution 
on each column of a Data Frame.
Can provide interpretation for p-value with "interpret=TRUE".
Modified from https://stackoverflow.com/questions/45226457/r-loop-performing-ks-tests-across-data-frame-stored-in-matrix
}
\examples{
colKS(dataframe, interpret=TRUE)

}
\keyword{KS}
\keyword{Kolmogorov-Smirnov}
\keyword{data.frame}
\keyword{normalty}
