% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domtools.r
\name{colMedian}
\alias{colMedian}
\title{Median for Data Frames}
\usage{
colMedian(x)
}
\description{
Function to calculate Median for each column of a Data Frame.
Modified from https://stackoverflow.com/questions/37806387/r-calculate-standard-deviation-in-cols-in-a-data-frame-despite-of-na-values
}
\examples{
colMedian(dataframe)

}
\keyword{Median}
\keyword{data.frame}
