% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domtools.r
\name{colSD}
\alias{colSD}
\title{Standard Deviation for Data Frames}
\usage{
colSD(x)
}
\description{
Function to calculate Standard Deviation for each column of a Data Frame.
Can provide interpretation for p-value with "interpret=TRUE".
Modified from https://stackoverflow.com/questions/37806387/r-calculate-standard-deviation-in-cols-in-a-data-frame-despite-of-na-values
}
\examples{
colSD(dataframe

}
\keyword{Deviation}
\keyword{Standard}
\keyword{data.frame}
\keyword{sd}
\keyword{std}
\keyword{stdev}
