% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domtools.r
\name{colShapiro}
\alias{colShapiro}
\title{Shapiro-Wilk Test for Data Frames}
\usage{
colShapiro(x, interpret = FALSE)
}
\description{
Function to perform Shapiro-Wilk test for each column of a Data Frame.
Can provide interpretation for p-value with "interpret=TRUE".
Modified from https://stackoverflow.com/questions/45226457/r-loop-performing-ks-tests-across-data-frame-stored-in-matrix
}
\examples{
colShapiro(dataframe, interpret=TRUE)

}
\keyword{Shapiro}
\keyword{Shapiro-Wilk}
\keyword{data.frame}
\keyword{normality}
