% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domtools.r
\name{firstup}
\alias{firstup}
\title{Capitalise First Letter Function}
\usage{
firstup(x)
}
\description{
Function to capitalise the first letter of a character string.
From https://stackoverflow.com/questions/18509527/first-letter-to-upper-case/18509816
}
\examples{
firstup("rStudio")

}
\keyword{capitalise}
\keyword{character}
\keyword{letter}
\keyword{string}
