% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domtools.r
\name{input}
\alias{input}
\title{Input Prompt Function}
\usage{
input(text = "", lines = 1)
}
\description{
This function creates an input prompt with customisable text.
Multiple inputs are possible with "lines" option".
}
\examples{
input("Press [Enter] to Continue")
name = input("Enter your name: ", lines=2)
}
\keyword{input}
