% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domtools.r
\name{modes}
\alias{modes}
\title{Mode Average Function}
\usage{
modes(x)
}
\description{
Function to calculate the modal average.
}
\examples{
modes(sample(1:10, 10, replace=TRUE))

}
\keyword{average}
\keyword{mode}
