% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domtools.r
\name{newline}
\alias{newline}
\title{Newline Function}
\usage{
newline()
}
\description{
This function starts a new line on the terminal.
Can also be used to print a blank line.
Run makeActiveBinding("newline", newline, baseenv())
to be able to type "newline
}
\examples{
newline()
}
\keyword{newline}
