% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domtools.r
\name{nl_fun}
\alias{nl_fun}
\title{Newline Function}
\usage{
nl_fun()
}
\description{
This function starts a new line on the terminal.
Can also be used to print a blank line.
}
\examples{
nl_fun()
newline
}
\keyword{newline}
